+++ docs/changelog.md
@@ -1,5 +1,13 @@
 # Full changelog
 
+## Upcoming version
+
+- Mod's translation was moved to the separated folder `/locale/<localeCode>` and not requires the suffix ~~`.<localeCode>.json`~~
+- Old format content packs trigs an "unsafe" warning when they are loaded. (yellow only for debug mode)
+- Content pack in the new format (1.3) are safer.
+- Updated content pack definition format to version 1.3
+- Redone content and content pack loader
+
 ## 0.12.0 beta "Go out"
 
 Released **22nd May 2020**